some frontend code refactoring
