+++ packages/defiscan-frontend/src/utils/exportMarkdown.ts
@@ -1,6 +1,6 @@
 import type { CompiledReview } from '../types'
 import { formatUsdValue, stripChainPrefix } from './format'
-import { getKeyFindings, type KeyFinding } from './narrative'
+import { getKeyFindings, type KeyFinding } from './keyFindings'
 
 const HUMAN_ADMIN_TYPES = new Set([
   'EOA',
@@ -63,7 +63,9 @@ export function reviewToMarkdown(review: CompiledReview): string {
   } else {
     const totalFunds = funds.reduce(
       (sum, f) =>
-        sum + (f.balances?.totalUsdValue ?? 0) + (f.positions?.totalUsdValue ?? 0),
+        sum +
+        (f.balances?.totalUsdValue ?? 0) +
+        (f.positions?.totalUsdValue ?? 0),
       0,
     )
     lines.push(
@@ -156,9 +158,7 @@ export function reviewToMarkdown(review: CompiledReview): string {
             fn.directFundsUsd > 0
               ? ` (${formatUsdValue(fn.directFundsUsd)} at risk)`
               : ''
-          lines.push(
-            `- \`${fn.contractName}.${fn.functionName}()\`${fundsStr}`,
-          )
+          lines.push(`- \`${fn.contractName}.${fn.functionName}()\`${fundsStr}`)
         }
       }
       lines.push('')
@@ -224,7 +224,9 @@ export function reviewToMarkdown(review: CompiledReview): string {
   }
 
   lines.push('---')
-  lines.push(`*Generated by DeFiScan on ${new Date().toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })}*`)
+  lines.push(
+    `*Generated by DeFiScan on ${new Date().toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })}*`,
+  )
   lines.push('')
 
   return lines.join('\n')
